+++ crates/rpc/src/rpc.rs
@@ -2,15 +2,16 @@ pub mod auth;
 mod conn;
 mod extension;
 mod llm;
+mod message_stream;
 mod notification;
 mod peer;
-pub mod proto;
 
 pub use conn::Connection;
 pub use extension::*;
 pub use llm::*;
 pub use notification::*;
 pub use peer::*;
+pub use proto;
 pub use proto::{Receipt, TypedEnvelope, error::*};
 mod macros;
